+++ runtime/tsc_runtime.h
@@ -318,6 +318,7 @@ double tsc_date_get_local_part(const tsc_date_t* d, int part);
 double tsc_date_get_timezone_offset(const tsc_date_t* d);
 tsc_str_t* tsc_date_to_iso_string(const tsc_date_t* d);
 tsc_str_t* tsc_date_to_utc_string(const tsc_date_t* d);
+tsc_str_t* tsc_date_to_date_string(const tsc_date_t* d);
 tsc_str_t* tsc_date_to_string(const tsc_date_t* d);
 
 /* ------------- Error ------------- */